enemies now have lasers and they do appropriate damage to the player (they used to do 2 damage per shot now it's just one damage per shot)

+++ Assets/Scripts/UIManager.cs
@@ -29,9 +29,9 @@ public class UIManager : MonoBehaviour
         _scoreText.text = "Score: " + playerScore;
     }
 
-    public void UpdateLives(int currentLives)
+    public void UpdateLives(float currentLives)
     {
-        _livesImg.sprite = _livesSprites[currentLives];
+        _livesImg.sprite = _livesSprites[(int) currentLives];
         if (currentLives <= 0)
         {
             _gameManager.GameOver();
